feat(geometry): add GeometryCache and FrameScratch for optimized rendering
