+++ src/test/GeometryCache.test.ts
@@ -1,5 +1,5 @@
 import { MotionType, PathSegment } from '../visualizer/types';
-import { FrameScratch, GeometryCache } from '../webview/geometryCache';
+import { GeometryCache } from '../webview/GeometryCache';
 import { StyleBucket } from '../webview/renderBuckets';
 
 describe('GeometryCache', () => {
@@ -111,24 +111,5 @@ describe('GeometryCache', () => {
     // Midpoint absolute index = segmentStart + floor(5/2) = 0 + 2 = 2.
     expect(cache.segmentMidpoint[0]).toBe(2);
     expect(cache.worldPoints[6]).toBe(2); // x of point index 2
-  });
-});
-
-describe('FrameScratch', () => {
-  it('allocates buffers sized to a cache', () => {
-    const cache = GeometryCache.build([
-      {
-        type: MotionType.FEED,
-        points: [
-          { x: 0, y: 0, z: 0 },
-          { x: 1, y: 0, z: 0 },
-        ],
-      },
-    ]);
-    const scratch = FrameScratch.forCache(cache);
-    expect(scratch.screen.length).toBe(cache.pointCount * 2);
-    expect(scratch.pointDepth.length).toBe(cache.pointCount);
-    expect(scratch.segmentDepth.length).toBe(cache.segmentCount);
-    expect(scratch.sortedSegments.length).toBe(cache.segmentCount);
   });
 });
